+++ CHANGES.md
@@ -1,6 +1,7 @@
 ## Changes in 0.10.4
 
 * Reformatted code base using [black](https://black.readthedocs.io/).
+* Updated copyright headers
 
 ## Changes in 0.10.3